fix delete exam

+++ controller/src/main/java/com/nicico/training/controller/TestQuestionRestController.java
@@ -127,7 +127,7 @@ public class TestQuestionRestController {
         try {
             TestQuestion testQuestion = testQuestionService.getById(id);
             // delete exam from ELS
-            if (testQuestion.getOnlineFinalExamStatus()) { // ارسال شده به آنلاین
+            if (testQuestion.getOnlineFinalExamStatus()!=null && testQuestion.getOnlineFinalExamStatus()) { // ارسال شده به آنلاین
                 baseResponse = elsClient.deleteExamFromEls(testQuestion.getId(), testQuestion.getTestQuestionType());
                 if (baseResponse.getStatus() != 200) {
                     return new ResponseEntity<>(
@@ -135,6 +135,7 @@ public class TestQuestionRestController {
                 }
 
             }
+
             testQuestionService.delete(id);
             // delete exam from Database
             return new ResponseEntity<>(baseResponse, HttpStatus.OK);
